Auto scale config for multi-node training
Summary:
Lightning trainer set max step to cfg.SOLVER.MAX_ITER. However, this is the max iteration for all nodes, in multi-node training, we need to scale it down, as well as eval period and other configs.
This diff calls `auto_scale_world_size` before passing the config to trainer.

Differential Revision: D28140877

fbshipit-source-id: 090bdb1faf769d3050b8a102db1a19bd28c16286