support building d2go transform list from a list of strings instead of just config
Summary:
for sythetic data, we want to enable having different transforms for different dataloaders.

To do that, we need to be able to construct different kinds of transforms.

This means that using the cfg's hard-coded location isn't convenient - we'd have to edit the cfg during run time and call the build function multiple times

Differential Revision: D32486576

fbshipit-source-id: 767b63c5c787e31a67dbf8710ab9bab84a0651db