enable overloading get_data_loader_vis_wrapper
Summary:
Currently the lightning task rely on the default runner for the vis wrapper logic. This does not allow to overload the get_data_loader_vis_wrapper is subclasses of the lightning task class.
This diff fixes this issue and properly take the vis wrapper given by the overloaded get_data_loader_vis_wrapper functions in the runner.

Reviewed By: zhanghang1989

Differential Revision: D33190410

fbshipit-source-id: 48cb3a8fa4b11df41d025d115d21002991549ced