Fix Metal optimized models' augment with bundled inputs
Summary:
Pull Request resolved: https://github.com/facebookresearch/d2go/pull/216

Sanity check after augment with bundled inputs fails unless tensor is moved to the correct backend.

Fix warning where "-metal" or "-vulkan" is not correctly removed from the string.

Temporary fix: Remove the call to augment with bundled inputs, because Metal backend for iOS GPU is not available on devserver. The true fix to unblock bundled inputs will be to add an input preformatting step op into metal models to convert input to Metal tensors (and no-op if already a metal tensor). This is outside the scope of this diff.

Reviewed By: ymao1993

Differential Revision: D35574266

fbshipit-source-id: 9f7b5c72dff2e3cf0eddf871379b079a1ec658ff