Move setup_after_lightning to ligthing_train_net level. Remove circular dependency and make robust to overriding by users
Summary:
Pull Request resolved: https://github.com/facebookresearch/d2go/pull/235

*This diff is part of a stack which has the goal of "buckifying" D2 (https://github.com/facebookresearch/d2go/commit/87374efb134e539090e0b5c476809dc35bf6aedb)Go core and enabling autodeps and other tooling. The last diff in the stack introduces the TARGETS. The diffs earlier in the stack are resolving circular dependencies and other issues which prevent the buckification from occurring.*

Trying to break a circular dependency between runner <-> setup, I saw that the setup function, which is a hook from lightningmodule is importing and running setup_after_lightning_launch.

This is a possible cause of issues, because I have seen users in the codebase overriding this method without calling setup_after_lightning_launch, which is probably not what we intend.

This diff moves the use of setup_after_lightning_launch to the same place as the regular train_net, in the binary before the training or eval happen.

Reviewed By: tglik

Differential Revision: D35966944

fbshipit-source-id: 4460daa13c46b46dc0432480298ef0689cca8af1