Hardcode top-level directory, remove import for resource access
Summary:
Pull Request resolved: https://github.com/facebookresearch/d2go/pull/229

*This diff is part of a stack which has the goal of "buckifying" D2 (https://github.com/facebookresearch/d2go/commit/87374efb134e539090e0b5c476809dc35bf6aedb)Go core and enabling autodeps and other tooling. The last diff in the stack introduces the TARGETS. The diffs earlier in the stack are resolving circular dependencies and other issues which prevent the buckification from occurring.*

This diff changes the import paths being rerouted in the d2go package. Instead of local-importing the package, which creates a buck circular dependency, we can hardcode the top-level path. We know it is fixed because we are defining it with base_module.

For the other packages it can remain as-is.

Also add some dependencies annotations in preparation for the buckification, with manual.

Differential Revision: D35928513

fbshipit-source-id: 1d20bdf03a6c96ed96effe9ab1d641fe164b7f0f