Add CallFuncWithNameAndJsonFile
Summary:
Pull Request resolved: https://github.com/facebookresearch/d2go/pull/288

Add a new call function to handle adhoc datasets that needs to be called with dataset_name. This is mainly to handle the incompatibility of `extend_coco_load` function with AdhocDataset

For example,
* if we choose to register `extend_coco_load` function as CallFuncWithJsonFile, AdhocDataset will register the same function, while failed to update the `dataset_name` argument passed to `extended_coco_load`.

* If we choose to register `extend_coco_load` function normally, we couldn't use the `LOAD_KWARGS` that specifies additional fields like `image_direct_copy_keys` when register AdhocDatasets

Therefore adding the new function allows to update register function with specifying new json as well as new dataset_name in the same time.

Reviewed By: sstsai-adl

Differential Revision: D37100742

fbshipit-source-id: 93f4a7ac23f95812aca80c40872b15f92d449da0