support a layer that saves outputs
Summary:
This diff adds a layer `CachedLayer` which is meant to be used with dynamic mixin. This layer runs the original module and clones the output into a dictionary provided by the user.

The main use case is in distillation where we dynamically mixin these layers to the layers that the user wants to compute various losses.

See subsequent diffs to get integration with distillation.

Differential Revision: https://www.internalfb.com/diff/D40285573?entry_point=27

fbshipit-source-id: a4a588278f47946b89ada1a0a420a14c2b1d237c