Add an option to specify the period of metric gathering and writing in Trainer
Summary:
Pull Request resolved: https://github.com/facebookresearch/d2go/pull/469

X-link: https://github.com/facebookresearch/detectron2/pull/4785

Add an option to specify the period of metric gathering and writing in Trainer.

This feature is needed to optimize training speed for large-scale training jobs like generative AI. The reason is that the all_gather call in metric writing at every iteration is time-consuming when hundreds of gpus are used. This takes ~10% of the total training time. With this feature we can set the metric writing period as the same as cfg.WRITER_PERIOD=20 to reduce training time while still keeping metric logging the same to users

Differential Revision: https://www.internalfb.com/diff/D43098985?entry_point=27

fbshipit-source-id: d8710f48abe8cac7098a00eff630fd9adf8184db