Add FCOS quantization support
Summary:
Pull Request resolved: https://github.com/facebookresearch/d2go/pull/480

As titled, this diff is a follow up of D38102296, adding FCOS quantization support.

A few items:
1. Add the FCOSInferenceWrapper
2. Add `prepare_for_export` and `prepare_for_quant` in FCOS.

NOTE: To avoid changing the adoption of `CycleBatchNormList` as mentioned in the previous diff, I have to add a hacky solution in the `prepare_for_quant` function. Specifically, flatten the one-element CycleBatchNormList to a BatchNorm2d.

Differential Revision: D43522795

fbshipit-source-id: 8772908f14f3fe44b9404187df90136683252642