Resolve relative import for modeldef
Summary:
There is an issue with the relative import in the `__init__` file of modeldef that causes tests on GitHub CI to fail.
Specifically, the `FBNetV2ModelArch` is not correctly populated.
The internal CI does not detect such failures because we use the buck build system.
This diff fixes it.

Pull Request resolved: https://github.com/facebookresearch/d2go/pull/547

Reviewed By: patricksnape

Differential Revision: D46177424

fbshipit-source-id: 06b23b9b221c990cd15a2debff6def8cfb99743b