Utils part 1: create a separate buck target
Summary:
Pull Request resolved: https://github.com/facebookresearch/d2go/pull/551

The `utils` dir is broken down into two steps to deal with circular dependencies while keeping the diffs atomic. This diff creates TARGETS for the dirs. `utils`(except `demo_predictor.py`) and `utils/fb`. The TARGETS for `utils/testing` and `utils/demo_predictor.py` are introduced in down the stack in the diff D46096376.

Reviewed By: tglik

Differential Revision: D45912077

fbshipit-source-id: fb01969c5f5df97de8afaa24bee8492591059b4d